Removed time variables, added variable millsOfSecond. Updated unit test (deleted explicit data type conversion)

+++ src/test/java/ru/itpark/DepositCalculatorServiceTest.java
@@ -7,7 +7,7 @@ class DepositCalculatorServiceTest {
     @org.junit.jupiter.api.Test
     void depositCalculate() {
         {
-            double sumDeposit = (double) 100_000;
+            double sumDeposit = 100_000;
             int period = 12;
             double percentDeposit = 10;
             String dateStart = "21.10.2018";
@@ -18,7 +18,7 @@ class DepositCalculatorServiceTest {
             assertEquals(110_000, deposit);
         }
         {
-            double sumDeposit = (double) 150_000;
+            double sumDeposit = 150_000;
             int period = 14;
             double percentDeposit = 8;
             String dateStart = "21.10.2018";
@@ -29,7 +29,7 @@ class DepositCalculatorServiceTest {
             assertEquals(164_005, deposit);
         }
         {
-            double sumDeposit = (double) 700_000;
+            double sumDeposit = 700_000;
             int period = 24;
             double percentDeposit = 9.2;
             String dateStart = "21.10.2018";
